requirements.txt
Django==5.2
asgiref==3.8.1
sqlparse==0.5.3
tzdata==2025.2
dj-database-url==2.1.0
python-dotenv==1.0.1
psycopg2-binary==2.9.9
django-cors-headers==4.3.1
djangorestframework==3.16.0
requests==2.31.0
